moved serviceconfig to WiseTorrent.UI to remove cyclic dependencies
