+++ WiseTorrent.Utilities/WiseTorrent.Utilities.csproj
@@ -7,15 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.7" />
   </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\WiseTorrent.Parsing\WiseTorrent.Parsing.csproj" />
-    <ProjectReference Include="..\WiseTorrent.Peers\WiseTorrent.Peers.csproj" />
-    <ProjectReference Include="..\WiseTorrent.Pieces\WiseTorrent.Pieces.csproj" />
-    <ProjectReference Include="..\WiseTorrent.Storage\WiseTorrent.Storage.csproj" />
-    <ProjectReference Include="..\WiseTorrent.Trackers\WiseTorrent.Trackers.csproj" />
-  </ItemGroup>
-
 </Project>
